ToUnicode.pm (_convert_commands): converted commands must be preceded by a not-consumed non-backslash   (else "\\ " will convert "\ " as a control space), and possibly followed by a not-consumed right brace   (else "\\}" would not be found), and check for beginning-of-string separately.
(_convert_accents): for the \"a case, the argument must be a single
character (\w), not two (\w{1,2}) as for all the others (to support the \t
tie-after accent). Allow and ignore whitespace after the accent command.

Tables.pm (COMMANDS): convert \<tab> and some of the spacing
commands: \thinspace \medspace \thickspace and their control-symbol and
\neg counterparts. Also \quad and \qquad.